feat: 구매자  컴포넌트 분리
