+++ src/components/PhotoBuyer/ExchangeCard.jsx
@@ -1,9 +1,13 @@
+"use client";
+
 import React, { useEffect, useState } from "react";
 import { Title } from "../ui/Title";
 import GradeTag from "../tag/GradeTag";
+import MyCardModal from "../CardSeller/MyCardModal";
 
 function ExchangeCard({ desiredDescription, cardGradeId, cardGenreId }) {
   const [isMd, setIsMd] = useState(false);
+  const [isOpen, setIsOpen] = useState(false);
 
   useEffect(() => {
     function onResize() {
@@ -14,7 +18,6 @@ function ExchangeCard({ desiredDescription, cardGradeId, cardGenreId }) {
     return () => window.removeEventListener("resize", onResize);
   }, []);
 
-  // 장르 ID → 장르명 매핑
   const genreMap = {
     1: "여행",
     2: "풍경",
@@ -29,15 +32,18 @@ function ExchangeCard({ desiredDescription, cardGradeId, cardGenreId }) {
         title="교환 희망 정보"
         buttonText={isMd ? "포토카드 교환하기" : undefined}
         font="titleLg_Noto"
+        onButtonClick={() => setIsOpen(true)}
       />
       <p className="mt-2 text-sm text-gray-700">{desiredDescription}</p>
       <p className="mt-1 text-sm text-gray-500">
         카드 등급: <GradeTag grade={cardGradeId} />
       </p>
       <p className="mt-1 text-sm text-gray-500">카드 장르: {genreText}</p>
+
+      {/* 교환 모달 */}
+      <MyCardModal isOpen={isOpen} onClose={() => setIsOpen(false)} />
     </div>
   );
 }
 
 export default ExchangeCard;
-
