+++ src/components/PhotoBuyer/PhotoBuyerSection.jsx
@@ -6,7 +6,6 @@ import CardBuyer from "../CardBuyer/CardBuyer";
 import { Title } from "../ui/Title";
 import { storeService } from "@/lib/api/api-store";
 
-
 export default function PhotoBuyerSection({ photo }) {
   const genreMap = {
     1: "여행",
@@ -14,7 +13,10 @@ export default function PhotoBuyerSection({ photo }) {
     3: "인물",
     4: "사물",
   };
+
   const [isLoading, setIsLoading] = useState(false);
+  const [remaining, setRemaining] = useState(photo.saleQuantity);
+
   const handleBuy = async (quantity) => {
     if (quantity <= 0) {
       alert("구매 수량을 선택해주세요.");
@@ -24,11 +26,10 @@ export default function PhotoBuyerSection({ photo }) {
     setIsLoading(true);
 
     try {
-      const purchaseResult = await storeService.purchaseCard(
-        photo.id,
-        quantity
-      );
-      alert(`구매 성공! ${purchaseResult.purchasedQuantity}장 구매되었습니다.`);
+      const result = await storeService.purchaseCard(photo.id, quantity);
+      alert(`구매 성공! ${result.purchasedQuantity}장 구매되었습니다.`);
+
+      setRemaining((prev) => prev - quantity);
     } catch (error) {
       alert(`구매 실패: ${error.message}`);
     } finally {
@@ -38,10 +39,10 @@ export default function PhotoBuyerSection({ photo }) {
 
   const imageUrl = photo?.photoCard?.imageUrl || example;
   const name = photo?.photoCard?.name || "제목 없음";
+
   return (
     <section className="w-full">
       <Title title={name} font="titleLg_Noto" />
-
       <div className="md:flex">
         <div className="w-full h-[260px] lg:h-[720px] overflow-hidden rounded-md mb-4 relative">
           <img
@@ -53,15 +54,16 @@ export default function PhotoBuyerSection({ photo }) {
         </div>
 
         <CardBuyer
+          cardId={photo.id}
           tier={photo?.photoCard?.gradeId}
           subLabel={genreMap[photo?.photoCard?.genreId]}
           creator={photo?.photoCard?.creator?.nickname || "익명"}
           description={photo?.photoCard?.description}
           pricePerCard={photo?.price}
-          remaining={photo?.saleQuantity}
+          remaining={remaining}
           total={photo?.photoCard?.totalQuantity}
-          onBuy={handleBuy}
-           isLoading={isLoading}
+          onSuccess={handleBuy}
+          isLoading={isLoading}
         />
       </div>
     </section>
